+++ sdmx-dl-api/src/main/java/internal/sdmxdl/web/spi/FailsafeSdmxWebDriver.java
@@ -1,17 +1,17 @@
 /*
  * Copyright 2019 National Bank of Belgium
- * 
- * Licensed under the EUPL, Version 1.1 or - as soon they will be approved 
+ *
+ * Licensed under the EUPL, Version 1.1 or - as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
- * 
+ *
  * http://ec.europa.eu/idabc/eupl
- * 
- * Unless required by applicable law or agreed to in writing, software 
+ *
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package internal.sdmxdl.web.spi;
@@ -31,7 +31,6 @@ import java.util.function.Consumer;
 import java.util.logging.Level;
 
 /**
- *
  * @author Philippe Charles
  */
 @lombok.extern.java.Log
@@ -42,8 +41,8 @@ public final class FailsafeSdmxWebDriver implements SdmxWebDriver {
         return obj instanceof FailsafeSdmxWebDriver
                 ? obj
                 : new FailsafeSdmxWebDriver(obj,
-                        FailsafeSdmxWebDriver::logUnexpectedError,
-                        FailsafeSdmxWebDriver::logUnexpectedNull);
+                FailsafeSdmxWebDriver::logUnexpectedError,
+                FailsafeSdmxWebDriver::logUnexpectedNull);
     }
 
     @lombok.NonNull
@@ -81,6 +80,16 @@ public final class FailsafeSdmxWebDriver implements SdmxWebDriver {
         } catch (RuntimeException ex) {
             onUnexpectedError.accept("Unexpected exception while getting rank", ex);
             return UNKNOWN;
+        }
+    }
+
+    @Override
+    public boolean isAvailable() {
+        try {
+            return delegate.isAvailable();
+        } catch (RuntimeException ex) {
+            onUnexpectedError.accept("Unexpected exception while getting availability", ex);
+            return false;
         }
     }