+++ sdmx-dl-api/src/test/java/sdmxdl/web/SdmxWebManagerTest.java
@@ -31,6 +31,8 @@ import java.time.Duration;
 import java.util.*;
 
 import static org.assertj.core.api.Assertions.*;
+import static sdmxdl.web.spi.SdmxWebDriver.NATIVE_RANK;
+import static sdmxdl.web.spi.SdmxWebDriver.WRAPPED_RANK;
 
 /**
  * @author Philippe Charles
@@ -75,7 +77,8 @@ public class SdmxWebManagerTest {
         SdmxWebDriver sdmx21 = MockedWebDriver
                 .builder()
                 .name("sdmx21")
-                .rank(SdmxWebDriver.WRAPPED_RANK)
+                .rank(WRAPPED_RANK)
+                .available(true)
                 .source(nbb)
                 .source(ecb)
                 .build();
@@ -141,11 +144,11 @@ public class SdmxWebManagerTest {
     public void testGetDefaultSources() {
         SdmxWebSource source1a = SdmxWebSource.builder().name("s1").driver("dX").endpointOf("http://abc").build();
         SdmxWebSource source2 = SdmxWebSource.builder().name("s2").driver("dX").endpointOf("http://abc").build();
-        SdmxWebDriver driverX = MockedWebDriver.builder().name("dX").rank(SdmxWebDriver.WRAPPED_RANK).source(source1a).source(source2).build();
+        SdmxWebDriver driverX = MockedWebDriver.builder().name("dX").rank(WRAPPED_RANK).available(true).source(source1a).source(source2).build();
 
         SdmxWebSource source1b = SdmxWebSource.builder().name("s1").driver("dY").endpointOf("http://xyz").build();
         SdmxWebSource source3 = SdmxWebSource.builder().name("s3").driver("dY").endpointOf("http://xyz").build();
-        SdmxWebDriver driverY = MockedWebDriver.builder().name("dY").rank(SdmxWebDriver.NATIVE_RANK).source(source1b).source(source3).build();
+        SdmxWebDriver driverY = MockedWebDriver.builder().name("dY").rank(NATIVE_RANK).available(true).source(source1b).source(source3).build();
 
         assertThat(SdmxWebManager.builder().driver(driverX).driver(driverY).build().getDefaultSources())
                 .containsExactly(source1a, source2, source3);
@@ -170,7 +173,8 @@ public class SdmxWebManagerTest {
         SdmxWebDriver driver1 = MockedWebDriver
                 .builder()
                 .name("d1")
-                .rank(SdmxWebDriver.WRAPPED_RANK)
+                .rank(WRAPPED_RANK)
+                .available(true)
                 .repo(asURL("http://abc"), sample)
                 .source(SdmxWebSource.builder().name("source").driver("d1").dialect("azerty").endpointOf("http://abc").build())
                 .build();
@@ -178,7 +182,8 @@ public class SdmxWebManagerTest {
         SdmxWebDriver driver2 = MockedWebDriver
                 .builder()
                 .name("d2")
-                .rank(SdmxWebDriver.NATIVE_RANK)
+                .rank(NATIVE_RANK)
+                .available(true)
                 .repo(asURL("http://xyz"), sample)
                 .source(SdmxWebSource.builder().name("source").driver("d2").dialect("azerty").endpointOf("http://xyz").build())
                 .build();
@@ -250,6 +255,7 @@ public class SdmxWebManagerTest {
             .builder()
             .name("repoDriver")
             .rank(0)
+            .available(true)
             .repo(asURL(sample), sample)
             .supportedProperty("someproperty")
             .source(sampleSource)
@@ -281,6 +287,9 @@ public class SdmxWebManagerTest {
 
         @lombok.Getter
         private final int rank;
+
+        @lombok.Getter
+        private final boolean available;
 
         @lombok.Singular
         private final Map<URL, SdmxRepository> repos;