+++ sdmx-dl-provider-ri/src/main/java/internal/sdmxdl/ri/web/drivers/FileDriver.java
@@ -43,6 +43,11 @@ public final class FileDriver implements SdmxWebDriver {
     }
 
     @Override
+    public boolean isAvailable() {
+        return true;
+    }
+
+    @Override
     public @NonNull SdmxWebConnection connect(@NonNull SdmxWebSource source, @NonNull SdmxWebContext context) throws IOException, IllegalArgumentException {
         Objects.requireNonNull(source, "source");
         Objects.requireNonNull(context, "context");